have only one tree active at a time
